Fix random chapter skipping
In case a chapter further than the starting chapter has a duration which
makes skipped_duration + duration < seek_positon that chapter would be
silently skipped. Lets go back to using a separate function to find the
position to fix that.

Additionally, bring back the guard to prevent landing in the final
couple of seconds of a track for more smoother chapter seek